Change signed indices' types to int64_t

+++ fenwick-tree.hpp
@@ -18,7 +18,7 @@ struct FenwickTree
         t = vector<T>(n, 0);
     }
 
-    T prefix_query(ptrdiff_t i)
+    T prefix_query(int64_t i)
     {
         i++;
         T x = 0;
@@ -30,12 +30,12 @@ struct FenwickTree
         return x;
     }
 
-    T range_query(ptrdiff_t i, ptrdiff_t j)
+    T range_query(int64_t i, int64_t j)
     {
         return prefix_query(j) - (i ? prefix_query(i - 1) : 0);
     }
 
-    void increment(ptrdiff_t i, T x)
+    void increment(int64_t i, T x)
     {
         i++;
         while ((size_t)i <= t.size())
@@ -45,7 +45,7 @@ struct FenwickTree
         }
     }
 
-    void set(ptrdiff_t i, T x)
+    void set(int64_t i, T x)
     {
         increment(i, x - range_query(i, i));
     }
@@ -64,23 +64,23 @@ struct FenwickTreeRURQ
         t[1] = FenwickTree<T>(n);
     }
 
-    T prefix_query(ptrdiff_t i)
+    T prefix_query(int64_t i)
     {
         return t[0].prefix_query(i) * (i + 1) - t[1].prefix_query(i);
     }
 
-    T range_query(ptrdiff_t i, ptrdiff_t j)
+    T range_query(int64_t i, int64_t j)
     {
         return prefix_query(j) - (i ? prefix_query(i - 1) : 0);
     }
 
-    void suffix_increment(ptrdiff_t i, T x)
+    void suffix_increment(int64_t i, T x)
     {
         t[0].increment(i, x);
         t[1].increment(i, x * i);
     }
 
-    void increment(ptrdiff_t i, ptrdiff_t j, T x)
+    void increment(int64_t i, int64_t j, T x)
     {
         suffix_increment(i, x);
         if ((size_t)j + 1 < t[0].t.size())
